requirements.txt
fastapi~=0.85.0
fastapi-utils
python-multipart
uvicorn[standard]~=0.18.3
gunicorn
sqlalchemy~=1.4.41
pydantic
docker~=6.0.0
jinja2
python_on_whales
ldap3~=2.9.1
python-dotenv~=0.21.0
kubernetes
pytest~=7.1.3
starlette~=0.20.4
